requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-environ==0.11.2
whitenoise==6.8.2

firebase-admin==7.1.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0

gunicorn==23.0.0
requests==2.32.3
python-dotenv==1.1.0
dj-database-url==2.2.0
python-decouple==3.8
psycopg[binary]==3.2.9
